Got a proof of concept down for the config file. (Its layout is only sufficient for a standard 2D board, however, I want a spec created that can handle triangular shaped boards, boards with cells scattered about etc..  I gave one "custom square setup example")
One particular neat type of chess variant I'd like to see is chess squares dealt out like cards on the form, (with random pieces or blanks) and you play it how they fall..

Test Board is now loading up from the config file

However, all that fabu config code needs to be moved from UIBoard to Board (and all the Board functions need to be modded to take XmlNodes)

All that Xml parsing crap needs to be moved to its own class (in Engine)

+++ trunk/source/Engine/Engine/Types/Boards/BoardDefinition.cs
@@ -7,8 +7,8 @@ namespace Engine.Types
 {
     public class BoardDef
     {
-        int _rows;
-        public int Rows
+        Int16 _rows;
+        public Int16 Rows
         {
             get
             {
@@ -20,8 +20,8 @@ namespace Engine.Types
             }
         }
 
-        int _cols;
-        public int Columns
+        Int16 _cols;
+        public Int16 Columns
         {
             get
             {
